Various fixes
1)

If output is piped into `less`, and `less` aborts, then the ensuing
EPIPE failures should not cause a panic. See Rust issue:

    https://github.com/rust-lang/rust/issues/46016

The workaround is to not use `print!()` or `println!()`, unfortunately.

2)

Before this change, if any of the child processes has terminated
pty-for-each has not really waited for the pid, but only for the closing
the terminal pipe. This change makes pty-for-each wait for both.

3)

Any non-zero exit status for one or more of the child processes should
cause a non-zero exit status from pty-for-each by default. This change
implements an exit status of 0xfe if all processes exited with a signal
0xff if all processes exited normally with a non-zero exit status, and
0xfd on mixed situations of the former two.